+++ src/main/kotlin/de/schwarzland/mavenup/service/MavenUpSettings.kt
@@ -58,7 +58,7 @@ class MavenUpSettings : PersistentStateComponent<MavenUpSettings.State> {
      */
     data class State(
         var jumpOnSingleClick: Boolean = false,
-        var selectLatestVersion: Boolean = true,
+        var selectLatestVersion: Boolean = false,
         var hideUnstableVersions: Boolean = false,
         var hiddenVersionQualifiers: String = "rc,beta,alpha,ea,milestone,preview,cr,nightly,snapshot",
         var ossIndexEnabled: Boolean = false,